fix training args mode

+++ mindone/transformers/models/llama/modeling_llama.py
@@ -1240,7 +1240,7 @@ class LlamaForSequenceClassification(LlamaPreTrainedModel):
                 else:
                     loss = self.loss_fct_regression(pooled_logits, labels)
             elif problem_type == 1:  #"single_label_classification"
-                loss = self.loss_fct_single_label_classification(pooled_logits.view(-1, self.num_labels), labels.view(-1))
+                loss = self.loss_fct_single_label_classification(pooled_logits.view(-1, self.num_labels), labels.view(-1).int())
             elif problem_type == 2:  #"multi_label_classification"
                 loss = self.loss_fct_multi_label_classification(pooled_logits, labels)
